Movement!
Movement is basically implemented, but with few minor bugs.

Need to make sure on mouse down doesn't pick up the button click.
Something was wrong with the sequence number checking so got rid of that for now.
Need to figure out something better to do with the selector when in movement mode.

Signed-off-by: Joseph Satterfield <jjs228@case.edu>